+++ releasenotes.md
@@ -44,6 +44,7 @@
 - Compiler segfault on global slice initialization with null[:0] #2404.
 - Use correct allocator in `replace`.
 - Regression: 1 character module names would create an error.
+- Compiler segfault with struct containing list of structs with an inline member #2416
 
 ### Stdlib changes
 - Add `==` to `Pair`, `Triple` and TzDateTime. Add print to `Pair` and `Triple`.